defined conditionally rendering when content is available and not available
