+++ lib/features/home/widgets/content.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class Content extends StatefulWidget {
+  const Content({super.key});
+
+  @override
+  State<Content> createState() => _ContentState();
+}
+
+class _ContentState extends State<Content> {
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
